tests aws credentials, fixes bucket tests
