+++ tests/load/filesystem/test_aws_credentials.py
@@ -19,7 +19,7 @@ def test_aws_credentials_resolved_from_default(environment: Dict[str, str]) -> N
     assert config.aws_access_key_id == 'fake_access_key'
     assert config.aws_secret_access_key == 'fake_secret_key'
     assert config.aws_session_token == 'fake_session_token'
-    assert config.aws_profile == 'default'
+    assert config.profile_name == 'default'
 
 
 @pytest.mark.skipif('s3' not in ALL_FILESYSTEM_DRIVERS, reason='s3 filesystem driver not configured')
@@ -30,10 +30,15 @@ def test_aws_credentials_from_boto3(environment: Dict[str, str]) -> None:
 
     import boto3
 
-    c = AwsCredentials(boto3.Session())
-    assert c.aws_profile == "default"
+    session = boto3.Session()
+
+    c = AwsCredentials(session)
+    assert c.profile_name == "default"
     assert c.aws_access_key_id == "fake_access_key"
+    assert c.region_name == session.region_name
+    assert c.profile_name == session.profile_name
     assert c.is_resolved()
+    assert not c.is_partial()
 
     c = AwsCredentials()
     c.parse_native_representation(boto3.Session())
