+++ tests/load/pipeline/utils.py
@@ -14,7 +14,7 @@ from dlt.pipeline.exceptions import SqlClientNotAvailable
 if TYPE_CHECKING:
     from dlt.destinations.filesystem.filesystem import FilesystemClient
 
-from tests.load.utils import ALL_DESTINATIONS, AWS_BUCKET, GCS_BUCKET, FILE_BUCKET
+from tests.load.utils import ALL_DESTINATIONS, AWS_BUCKET, GCS_BUCKET, FILE_BUCKET, ALL_BUCKETS
 
 
 @dataclass
@@ -57,6 +57,20 @@ class DestinationTestConfiguration:
         self.setup()
         pipeline = dlt.pipeline(pipeline_name=pipeline_name, destination=self.destination, staging=self.staging, dataset_name=dataset_name or pipeline_name, full_refresh=full_refresh)
         return pipeline
+
+
+# TODO: please merge this with destinations_configs
+_destinations_and_buckets = [(d, ) for d in ALL_DESTINATIONS] + [('filesystem', b) for b in ALL_BUCKETS]
+_param_ids = [':'.join(p) for p in _destinations_and_buckets]
+
+@pytest.fixture(scope='function', params=_destinations_and_buckets, ids=_param_ids)
+def any_destination(request) -> Iterator[Tuple[str, ...]]:  # type: ignore[no-untyped-def]
+    """Parametrized fixture iterating all destinations including filesystem configured for each test bucket"""
+    destination = request.param[0]
+    if len(request.param) > 1:
+        os.environ['DESTINATION__FILESYSTEM__BUCKET_URL'] = request.param[1]
+    yield destination
+
 
 def destinations_configs(
         default_non_staging_configs: bool = False,
