+++ tests/load/utils.py
@@ -28,12 +28,12 @@ from dlt.destinations.job_client_impl import SqlJobClientBase
 from tests.utils import ALL_DESTINATIONS
 
 # bucket urls
-AWS_BUCKET = dlt.config.get("tests.bucket_url_aws", str)
-GCS_BUCKET = dlt.config.get("tests.bucket_url_gcs", str)
+AWS_BUCKET = dlt.config.get("tests.bucket_url_s3", str)
+GCS_BUCKET = dlt.config.get("tests.bucket_url_gs", str)
 FILE_BUCKET = dlt.config.get("tests.bucket_url_file", str)
 MEMORY_BUCKET = dlt.config.get("tests.memory", str)
 
-ALL_FILESYSTEM_DRIVERS = dlt.config.get("ALL_FILESYSTEM_DRIVERS", list) or ["aws", "gcs", "file", "memory"]
+ALL_FILESYSTEM_DRIVERS = dlt.config.get("ALL_FILESYSTEM_DRIVERS", list) or ["s3", "gs", "file", "memory"]
 
 # Filter out buckets not in all filesystem drivers
 ALL_BUCKETS = [GCS_BUCKET, AWS_BUCKET, FILE_BUCKET, MEMORY_BUCKET]
